requirements.txt
transformers[pytorch]
transformers[flax]
dataclass-wizzard  
pyzmq
